Change the ping timeouts for less disconnects
This is an ongoing problem because the suit only sends pings in response to requests. In new versions it should send at regular intervals. Now, we can detect a connection within half a second, but a disconnection only within ~3 seconds. But this lets us not disconnect when people are spamming some effects.