Brain state: figuring out if Nodes are a necessary abstraction
Right now, plugins expose their apis in a sort of global fashion. Under the covers they may have multiple devices, and we may have the apis take device ids. But should these different nodes in the manifest be exposed in the Driver? should the driver be picking and choosing which nodes to activate?