A lot of work on the BodyGraph.
Brain dump: the process is like this. First, the hardware plugins register their regions as coordinates on bodyparts. Then, these are transformed into the most specific named subregion that the API currently supports. When a user of the API requests a haptic to play on a named subregion, we then look and find the best matching one present. For instance, if the API has very high granularity but the user requests a general region, it will find the devices that are part of that greater region, but still the most specific device available.