The Great Serial Refactoring of 2017
Refactored BoostSerialAdapter into two classes. Now the serial port is contained in its own SerialPort class, which encapsulates connecting to a port for the first time. It is now possible to send pings to all the ports on the system at once, and then wait for the responses to come in.