Settle for adding another std::function to callback for fake interfaces
It's really annoying to wire up the function pointers in the c interfaces. You have to static cast void*, which isn't that much of a problem, but then you need to make a member function which does the legwork. So I added another std::function to the callback<> interface, and now you can use a lambda like whatever_interface.handler.cpp_fn = [this](args) {...}

Unsure of performance implications if any.